functions/asset/initiate/requirements.txt
boto3
pyhumps
ujson
marshmallow-dataclass